# Conflicts:
#	src/jdk.jshell/share/classes/jdk/jshell/SourceCodeAnalysisImpl.java

+++ src/java.base/share/classes/java/io/CharArrayReader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1996, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1996, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -196,7 +196,7 @@ public class CharArrayReader extends Reader {
             if (n < 0) {
                 return 0;
             }
-            pos += n;
+            pos += (int) n;
             return n;
         }
     }
